+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.2] - 2021-06-15
+### Added
+- New sample degradation options (only affects input, similar to random warp): 
+  - Random noise (gaussian/laplace/poisson)
+  - Random blur (gaussian/motion)
+  - Random jpeg compression
+  - Random downsampling
+- New "warped" preview(s): Shows the input samples with any/all distortions.     
+
 ## [1.7.1] - 2021-06-15
 ### Added
 - New autobackup options: